--- e2e/wdio/commonjs/package.json
@@ -1,3 +1,0 @@
-{
-  "type": "commonjs"
-}
